Tidy up the file
