+++ CONTRIBUTING.md
@@ -17,7 +17,7 @@ Some tips:
   [CredWriteA](https://docs.microsoft.com/en-us/windows/win32/api/wincred/nf-wincred-credwritea).
 
 - To create a new API, *don't* edit the main library files themselves; these get
-  overwritten. Instead, edit `tool\win32\win32api.csv` and run
+  overwritten. Instead, edit `tool\win32\win32api.json` and run
   `tool\generate.cmd` to update the library files. This might seem a pain, but
   it gives protection against any changes to dart:ffi, since the typedef syntax
   can be easily recreated as necessary.
