+++ tool/win32/converter.dart
@@ -1,35 +1,5 @@
-import 'dart:convert';
-import 'dart:io';
-
+import 'model.dart';
 import 'win32types.dart';
-
-// THIS IS TERRIBLE CODE AND I NEED TO UNTANGLE IT
-
-final prototypes = <String, TypeDef>{};
-
-class TypeDef {
-  final String neutralApiName;
-  final List<String> prototype;
-
-  final String dllLibrary;
-  final String comment;
-
-  const TypeDef(
-      this.neutralApiName, this.dllLibrary, this.prototype, this.comment);
-
-  Map<String, dynamic> toJson() => <String, dynamic>{
-        'neutralApiName': neutralApiName,
-        'prototype': prototype.first,
-        'dllLibrary': dllLibrary,
-        'comment': comment
-      };
-
-  TypeDef.fromJson(Map<String, dynamic> json)
-      : neutralApiName = json['neutralApiName'] as String,
-        prototype = [json['prototype'] as String],
-        dllLibrary = json['dllLibrary'] as String,
-        comment = json['comment'] as String;
-}
 
 String dartFromFFI(String ffiType) {
   if (ffiType.startsWith('Pointer')) {
@@ -144,21 +114,6 @@ Win32Function loadFunction(String rawFunction) {
   final func = Win32Function(returnType, apiName, params);
   win32APIs.add(func);
   return func;
-}
-
-void saveJson(String filename) {
-  final file = File(filename);
-  file.writeAsStringSync(json.encode(prototypes).replaceAll(r'\\n', r'\n'));
-}
-
-void loadJson(String filename) {
-  final file = File(filename);
-  final fileContents = file.readAsStringSync().replaceAll(r'\n', r'\\n');
-
-  final decoded = json.decode(fileContents) as Map<String, dynamic>;
-  for (final api in decoded.keys) {
-    prototypes[api] = TypeDef.fromJson(decoded[api] as Map<String, dynamic>);
-  }
 }
 
 void parsePrototypes() {
